refactor: extract shared ScrollbarView, add --include CLI flag, and update docs
- Move ScrollbarView from views/table/scrollbar/ to views/scrollbar/ as a
  shared component reused by both HelpView and TableView
- Simplify ScrollbarView API: handle_mouse(event, max_position) and
  render(area, buf, theme, position, max_position, viewport_size)
- Add --include/-i CLI flag for config composition (repeatable, merged
  left-to-right after config-file include_files)
- Replace Config::TryFrom<Option<PathBuf>> with Config::load(path, includes)
- Narrow style.rs function signatures (clipboard_style takes &Clipboard,
  header_style takes &Table)
- Use Duration instead of raw milliseconds in DoubleClick
- Use Vec<usize> instead of Vec<u16> for LineItemMap item heights
- Fix typo in Reset command comment
- Update README with --include documentation

+++ src/views/table/scroll.rs
@@ -64,7 +64,7 @@ pub(super) fn previous_page(
 mod tests {
     use super::{next_page, previous_page, LineItemMap};
 
-    fn map(heights: Vec<u16>, visible: usize, first: usize) -> LineItemMap {
+    fn map(heights: Vec<usize>, visible: usize, first: usize) -> LineItemMap {
         LineItemMap::new(heights, visible, first)
     }
